+++ CHANGELOG.md
@@ -23,6 +23,7 @@ _ `_optional` subpackage for managing optional dependencies
 - Abstract `ContinuousNonlinearConstraint` class
 - `ContinuousCardinalityConstraint` class and corresponding uniform sampling mechanism
 - `register_hooks` utility enabling user-defined augmentation of arbitrary callables
+- `DiscreteCardinalityConstraint` class
 
 ### Changed
 - Passing an `Objective` to `Campaign` is now optional